+++ lib/runger_style/cops/multiline_method_arguments_line_breaks.rb
@@ -8,7 +8,6 @@ module RungerStyle # rubocop:disable Style/ClassAndModuleChildren
     MSG = 'Each argument in a multi-line method call must start on a separate line.'
 
     # rubocop:disable Metrics/PerceivedComplexity
-    # rubocop:disable Metrics/CyclomaticComplexity
     def on_send(node)
       if node.arguments? && multiline_method_call?(node)
         # When a method call uses keyword arguments without braces,
@@ -35,7 +34,6 @@ module RungerStyle # rubocop:disable Style/ClassAndModuleChildren
         end
       end
     end
-    # rubocop:enable Metrics/CyclomaticComplexity
     # rubocop:enable Metrics/PerceivedComplexity
 
     private